// GradesAppFront/src/app/models/created-asignatura.interface.ts
// Generated by https://quicktype.io

export class PostAsignaturaDTO {
  nombre: string;
  horas: number;
  idProfesor: string;
  descripcion: string;
  color: string;
  constructor(
    nombre: string,
    horas: number,
    idProfesor: string,
    descripcion: string,
    color: string
  ) {
    this.nombre = nombre;
    this.horas = horas;
    this.idProfesor = idProfesor;
    this.descripcion = descripcion;
    this.color = color;
  }
}